Fixing the movement of the player

+++ includes/header.h
@@ -43,8 +43,8 @@
 # define ABS(N) ((N<0)?(-N):(N))
 
 void	draw_map();
-void	dda(int X0, int Y0, int X1, int Y1);
-void	ft_square(int x, int y, int color, int size);
+void	dda(float X0, float Y0, float X1, float Y1);
+void	ft_square(float x, float y, int color, int size);
 void	player_init();
 int		deal_key();
 int		key_pressed(int keycode);
@@ -59,7 +59,7 @@ void	check_map(char *line, s_map *_map);
 void	fill_line(char *line, s_map *_map);
 void	map_manager(s_map *_map);
 void	ft_put_image(int x, int y, int color);
-void  	ft_dda(int X, int Y, int X1, int Y1);
+void  	ft_dda(float X, float Y, float X1, float Y1);
 void  	ft_3d_walls(float x, float y);
 
 /*
